Prevent array overflow when initializing registry
When collecting all enabled label definitions from the heat pump model's header
file, the corresponding registry IDs are collected in an array.
The problem is that the array can hold at most 32 elements but there are more
registry entries in the header that can be potentially added to the registry
array. Basically, this can result in an array overflow when more than 32
registry entries are enabled (uncommented) in the header file.