Fix serial write API to make sure all bytes are transmitted
We need to use the serial write API where the amount of bytes to be sent can be
chosen. That way, commands that contain a 0 byte are transmitted correctly.

When using the old API, the array that shall be sent is interpreted as a
C string. Therefore, any byte before a 0 byte in the array is seen as the last
byte to be sent.

Actual bug:
When querying register 0 from the heat pump, the third byte in the packet to be
sent is a 0 byte. Therefore, only the first two bytes (which are preceeding the
0 byte) are actually sent. The register ID (which is 0) and the two CRC bytes
are skipped.
The heat pump answers with 0x15 0xea which more or less means "query not
understood".