Add loop MQTT message and renamed topics
Add a MQTT message at beginnig and end of a loop to be able to use
it as trigger channel in e.g. openHAB when OneATTR is being defined.
It is meant as dispatcher message when dealing with calcultions of
e.g. the cop which is dependent on many different values.