Negative values as numeric type
Negative numbers will be added to json as numeric type instead of string. And only if minus sign is the first character, since converter might also return "---", which is not numeric. And there is no need to check all of the remaining characters in string, if alpha condition is met. Breaking out of the loop immediately.